mvc-spring-backend

+++ src/main/java/com/anhembiblog/anhembiblog/repository/PostRepo.java
@@ -6,6 +6,7 @@ import com.anhembiblog.anhembiblog.model.Post;
 import java.util.List;
 
 
-public interface PostRepo extends JpaRepository<Post, Integer>{
-    List<Post> findByAuthorId(int userId);
+public interface PostRepo extends JpaRepository<Post, Long>{
+    List<Post> findByAuthorId(Long userId);
+    
 }
